requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
Django==5.1.1
django-admin==2.0.2
django-cors-headers==4.4.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.2
excel-base==1.0.4
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
idna==3.8
isoweek==1.3.3
minio==7.2.8
outcome==1.3.0.post0
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
screen==1.0.1
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
TimeConvert==3.0.13
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
xlwt==1.3.0
